+++ askbot/models/__init__.py
@@ -272,7 +272,7 @@ def user_get_gravatar_url(self, size):
     """returns gravatar url
     """
     return GRAVATAR_TEMPLATE % {
-                'gravatar_url': askbot_settings.AVATAR_URL,
+                'gravatar_url': askbot_settings.GRAVATAR_BASE_URL,
                 'gravatar': self.gravatar,
                 'type': askbot_settings.GRAVATAR_TYPE,
                 'size': size,
@@ -356,7 +356,7 @@ def user_strip_email_signature(self, text):
     return text
 
 def _check_gravatar(gravatar):
-    gravatar_url = askbot_settings.AVATAR_URL+"/%s?d=404" % gravatar
+    gravatar_url = askbot_settings.GRAVATAR_BASE_URL + "/%s?d=404" % gravatar
     code = urllib.urlopen(gravatar_url).getcode()
     if urllib.urlopen(gravatar_url).getcode() != 404:
         return 'g' #gravatar